refactor and have single media config file
