+++ todo.md
@@ -1,0 +1,10 @@
+Generate manifests from the media_auth_config file too
+
+"Login" screen templates for interactive profiles - logins and clickthroughs and other things.
+Do some silly examples, too.
+
+Don't need "open" - absence of auth does this too. But still have max.
+
+Put w,h,d on all media.
+
+For the pattern of a static image with an image service, where both must be authed, the info.json of the image service is the static image's probe service.
